requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
langchain
langchain_community
langchain-text-splitters
transformers
torch
numpy
sentence-transformers
pypdf
chromadb
google-generativeai
google-genai
langchain-google-genai
pydantic
#python -m uvicorn app.main:app --reload